+++ internal/iam/command.go
@@ -1,6 +1,9 @@
 package iam
 
 import (
+	"fmt"
+	"github.com/confluentinc/cli/v4/pkg/ccloudv2"
+	"github.com/confluentinc/cli/v4/pkg/utils"
 	"github.com/spf13/cobra"
 
 	pcmd "github.com/confluentinc/cli/v4/pkg/cmd"
@@ -38,4 +41,50 @@ func New(cfg *config.Config, prerunner pcmd.PreRunner) *cobra.Command {
 	cmd.AddCommand(newUserCommand(cfg, prerunner))
 
 	return cmd
+}
+
+func AddResourceOwnerFlag(cmd *cobra.Command, cliCommand *pcmd.AuthenticatedCLICommand) {
+	items := []string{"user", "group-mapping", "service-account", "identity-pool"}
+	description := fmt.Sprintf("The resource ID of the principal who will be assigned resource owner on the "+
+		"created resource. Principal can be a %s.", utils.ArrayToCommaDelimitedString(items, "or"))
+	cmd.Flags().String("resource-owner", "", description)
+	pcmd.RegisterFlagCompletionFunc(cmd, "resource-owner", func(cmd *cobra.Command, args []string) []string {
+		if err := cmd.PersistentPreRunE(cmd, args); err != nil {
+			return nil
+		}
+
+		return AutocompleteResourceOwners(cliCommand.V2Client)
+	})
+}
+func AutocompleteResourceOwners(client *ccloudv2.Client) []string {
+	users, err := client.ListIamUsers()
+	if err != nil {
+		return nil
+	}
+	groupMappings, err := client.ListGroupMappings()
+	if err != nil {
+		return nil
+	}
+	serviceAccounts, err := client.ListIamServiceAccounts()
+	if err != nil {
+		return nil
+	}
+
+	suggestions := make([]string, len(users)+len(groupMappings)+len(serviceAccounts))
+	offset := 0
+	for i, user := range users {
+		description := user.GetFullName()
+		suggestions[i] = fmt.Sprintf("%s\t%s", user.GetId(), description)
+	}
+	offset += len(users)
+	for i, groupMapping := range groupMappings {
+		description := fmt.Sprintf("%s: %s", groupMapping.GetDisplayName(), groupMapping.GetDescription())
+		suggestions[i+offset] = fmt.Sprintf("%s\t%s", groupMapping.GetId(), description)
+	}
+	offset += len(groupMappings)
+	for i, serviceAccount := range serviceAccounts {
+		description := fmt.Sprintf("%s: %s", serviceAccount.GetDisplayName(), serviceAccount.GetDisplayName())
+		suggestions[i+offset] = fmt.Sprintf("%s\t%s", serviceAccount.GetId(), description)
+	}
+	return suggestions
 }
