+++ internal/iam/command_service_account_create.go
@@ -30,7 +30,7 @@ func (c *serviceAccountCommand) newCreateCommand() *cobra.Command {
 	}
 
 	cmd.Flags().String("description", "", "Description of the service account.")
-	pcmd.AddResourceOwnerFlag(cmd, c.AuthenticatedCLICommand)
+	AddResourceOwnerFlag(cmd, c.AuthenticatedCLICommand)
 	pcmd.AddContextFlag(cmd, c.CLICommand)
 	pcmd.AddOutputFlag(cmd)
